requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
aiofiles
python-multipart
python-jose